StoreThumbnail 컴포넌트 commmon 컴포넌트로 변경
